refactor: correct imports
there was nothing broken. but this is to make the code structure more uniform
