+++ src/modules/pets/stubs/index.ts
@@ -1,0 +1,1 @@
+export * from "./pets.stub";
